Upgrading from sbt 0.13.X to 1.2.X, moving src directories to new structure
I know this means that a lot of the code structure is eff:d up at this point, but it may still be worth it if we need to maintain this in the future.

The move from sbt 0.13 to 1.2 was to make it easier to override the publish settings. For some reason, the embedded build code and sbt 0.13 did not play nice. This should be more according to standard.